Third commit

+++ RestaurantApplication/src/main/java/com/npci/restaurantapp/services/RestaurantServices.java
@@ -5,10 +5,13 @@ package com.npci.restaurantapp.services;
 import java.util.List;
 import java.util.Optional;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+
 
 import com.npci.restaurantapp.dao.CommentDao;
 import com.npci.restaurantapp.dao.FoodItemDao;
@@ -16,6 +19,11 @@ import com.npci.restaurantapp.dao.RestaurantDao;
 import com.npci.restaurantapp.entity.Comment;
 import com.npci.restaurantapp.entity.FoodItem;
 import com.npci.restaurantapp.entity.Restaurant;
+import com.npci.restaurantapp.exceptions.CommentDescriptionException;
+import com.npci.restaurantapp.exceptions.FoodItemIdException;
+import com.npci.restaurantapp.exceptions.FoodItemNullException;
+import com.npci.restaurantapp.exceptions.RestaurantCommentException;
+import com.npci.restaurantapp.exceptions.RestaurantPincodeException;
 
 @Service
 public class RestaurantServices implements IRestaurantServices{
@@ -27,15 +35,27 @@ public class RestaurantServices implements IRestaurantServices{
 	@Autowired
 	CommentDao commentdao;
 
+	Logger LOGGER=LoggerFactory.getLogger(RestaurantServices.class);
+	
 	@Override
 	public Restaurant newRestaurant(Restaurant restaurant) {
+		String Pincocoderegx="^[0-9]{1,6}$";
+		Integer pincode = restaurant.getPincode();
+		if(String.valueOf(pincode).matches(Pincocoderegx)) {
+			throw new RestaurantPincodeException("Pincode Should be six Digit");
+		}
 		Restaurant newRestaurant = restaurantdao.save(restaurant);
+		LOGGER.info("Resturent detail:{}",newRestaurant);
 		return newRestaurant;
 	}
 
 	@Override
 	public FoodItem newFoodItem(FoodItem fooditem) {
+		if(fooditem.getItemType()==null) {
+			throw new FoodItemNullException("FoodItem type is null ");
+		}
 		FoodItem Fooditem = foodItemdao.save(fooditem);
+		LOGGER.info("FoodItem detail:{}",Fooditem);
 		return Fooditem;
 	}
 
@@ -43,20 +63,27 @@ public class RestaurantServices implements IRestaurantServices{
 
 	@Override
 	public String cutoutFoodUtem(Integer ItemID) {
+		LOGGER.info("ItemId :{}",ItemID);
+		Optional<FoodItem> FoodItem = foodItemdao.findById(ItemID);
+		if(!FoodItem.isPresent()) {
+			throw new FoodItemIdException("There is no Food Item with ItemID:"+ItemID);
+		}
+		
 		foodItemdao.deleteById(ItemID);
 		return "Food Item with ID:" + ItemID +"  deleted Successfully";
 	}
 
-//	@Override
-//	public FoodItem reviseFoodItem(FoodItem fooditem) {
-//		foodItemdao.fin
-//		return null;
-//	}
+
 	
 
 	public Comment newComment(@RequestBody Comment comment) {
-		Comment commentsave = commentdao.save(comment);
 		
+		if(comment.getComments()== null) {
+			throw new CommentDescriptionException("Comment Field is Empty");
+			
+		}
+		Comment commentsave = commentdao.save(comment);
+		LOGGER.info("Comment saved in table is:{}",commentsave);
 		return commentsave;
 		
 	}
@@ -69,7 +96,12 @@ public class RestaurantServices implements IRestaurantServices{
 
 	@Override
 	public List<Comment> Comments(Integer restID) {
+		LOGGER.info("Resturent ID :{}",restID);
 		List<Comment> findByRestaurantIdOrderByCIdAsc = commentdao.findByRestaurantIdOrderByCIdDesc(restID);
+		if(findByRestaurantIdOrderByCIdAsc.isEmpty()) {
+			throw new RestaurantCommentException("The is no comment for this restaurant id:"+restID);
+		}
+		LOGGER.info("Comment for Resturent ID {}",findByRestaurantIdOrderByCIdAsc);
 		return findByRestaurantIdOrderByCIdAsc;
 	}
